+++ coordinate/coordinatetest/attempt.go
@@ -1,4 +1,4 @@
-// Copyright 2015-2016 Diffeo, Inc.
+// Copyright 2015-2017 Diffeo, Inc.
 // This software is released under an MIT/X11 open source license.
 
 package coordinatetest
@@ -6,14 +6,12 @@ package coordinatetest
 import (
 	"github.com/diffeo/go-coordinate/cborrpc"
 	"github.com/diffeo/go-coordinate/coordinate"
-	"github.com/stretchr/testify/assert"
 	"reflect"
-	"testing"
 	"time"
 )
 
 // TestAttemptLifetime validates a basic attempt lifetime.
-func TestAttemptLifetime(t *testing.T) {
+func (s *Suite) TestAttemptLifetime() {
 	var (
 		err               error
 		data              map[string]interface{}
@@ -25,117 +23,117 @@ func TestAttemptLifetime(t *testing.T) {
 		WorkSpecName:  "spec",
 		WorkUnitName:  "a",
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
 
 	// The work unit should be "available"
-	sts.CheckUnitStatus(t, coordinate.AvailableUnit)
+	sts.CheckUnitStatus(s, coordinate.AvailableUnit)
 
 	// The work unit data should be defined but empty
-	DataEmpty(t, sts.WorkUnit)
+	s.DataEmpty(sts.WorkUnit)
 
 	// Get an attempt for it
-	attempt = sts.RequestOneAttempt(t)
+	attempt = sts.RequestOneAttempt(s)
 
 	// The work unit and attempt should both be "pending"
-	sts.CheckUnitStatus(t, coordinate.PendingUnit)
-	AttemptStatus(t, coordinate.Pending, attempt)
+	sts.CheckUnitStatus(s, coordinate.PendingUnit)
+	s.AttemptStatus(coordinate.Pending, attempt)
 
 	// The active attempt for the unit should match this
 	attempt2, err = sts.WorkUnit.ActiveAttempt()
-	if assert.NoError(t, err) {
-		AttemptMatches(t, attempt, attempt2)
+	if s.NoError(err) {
+		s.AttemptMatches(attempt, attempt2)
 	}
 
 	// There should be one active attempt for the worker and it should
 	// also match
 	attempts, err := sts.Worker.ActiveAttempts()
-	if assert.NoError(t, err) {
-		if assert.Len(t, attempts, 1) {
-			AttemptMatches(t, attempt, attempts[0])
+	if s.NoError(err) {
+		if s.Len(attempts, 1) {
+			s.AttemptMatches(attempt, attempts[0])
 		}
 	}
 
 	// The work unit data should (still) be defined but empty
-	DataEmpty(t, sts.WorkUnit)
+	s.DataEmpty(sts.WorkUnit)
 
 	// Now finish the attempt with some updated data
 	err = attempt.Finish(map[string]interface{}{
 		"outputs": []string{"yes"},
 	})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	// The unit and should report "finished"
-	sts.CheckUnitStatus(t, coordinate.FinishedUnit)
-	AttemptStatus(t, coordinate.Finished, attempt)
+	sts.CheckUnitStatus(s, coordinate.FinishedUnit)
+	s.AttemptStatus(coordinate.Finished, attempt)
 
 	// The attempt should still be the active attempt for the unit
 	attempt2, err = sts.WorkUnit.ActiveAttempt()
-	if assert.NoError(t, err) {
-		AttemptMatches(t, attempt, attempt2)
+	if s.NoError(err) {
+		s.AttemptMatches(attempt, attempt2)
 	}
 
 	// The attempt should not be in the active attempt list for the worker
 	attempts, err = sts.Worker.ActiveAttempts()
-	if assert.NoError(t, err) {
-		assert.Empty(t, attempts)
+	if s.NoError(err) {
+		s.Empty(attempts)
 	}
 
 	// Both the unit and the worker should have one archived attempt
 	attempts, err = sts.WorkUnit.Attempts()
-	if assert.NoError(t, err) {
-		if assert.Len(t, attempts, 1) {
-			AttemptMatches(t, attempt, attempts[0])
+	if s.NoError(err) {
+		if s.Len(attempts, 1) {
+			s.AttemptMatches(attempt, attempts[0])
 		}
 	}
 
 	attempts, err = sts.Worker.AllAttempts()
-	if assert.NoError(t, err) {
-		if assert.Len(t, attempts, 1) {
-			AttemptMatches(t, attempt, attempts[0])
+	if s.NoError(err) {
+		if s.Len(attempts, 1) {
+			s.AttemptMatches(attempt, attempts[0])
 		}
 	}
 
 	// This should have updated the visible work unit data too
 	data, err = sts.WorkUnit.Data()
-	if assert.NoError(t, err) {
-		assert.Len(t, data, 1)
-		if assert.Contains(t, data, "outputs") {
-			if assert.Len(t, data["outputs"], 1) {
-				assert.Equal(t, "yes", reflect.ValueOf(data["outputs"]).Index(0).Interface())
+	if s.NoError(err) {
+		s.Len(data, 1)
+		if s.Contains(data, "outputs") {
+			if s.Len(data["outputs"], 1) {
+				s.Equal("yes", reflect.ValueOf(data["outputs"]).Index(0).Interface())
 			}
 		}
 	}
 
 	// For bonus points, force-clear the active attempt
 	err = sts.WorkUnit.ClearActiveAttempt()
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	// This should have pushed the unit back to available
-	sts.CheckUnitStatus(t, coordinate.AvailableUnit)
+	sts.CheckUnitStatus(s, coordinate.AvailableUnit)
 
 	// This also should have reset the work unit data
-	DataEmpty(t, sts.WorkUnit)
+	s.DataEmpty(sts.WorkUnit)
 
 	// But, this should not have reset the historical attempts
 	attempts, err = sts.WorkUnit.Attempts()
-	if assert.NoError(t, err) {
-		if assert.Len(t, attempts, 1) {
-			AttemptMatches(t, attempt, attempts[0])
+	if s.NoError(err) {
+		if s.Len(attempts, 1) {
+			s.AttemptMatches(attempt, attempts[0])
 		}
 	}
 
 	attempts, err = sts.Worker.AllAttempts()
-	if assert.NoError(t, err) {
-		if assert.Len(t, attempts, 1) {
-			AttemptMatches(t, attempt, attempts[0])
+	if s.NoError(err) {
+		if s.Len(attempts, 1) {
+			s.AttemptMatches(attempt, attempts[0])
 		}
 	}
 }
 
 // TestAttemptMetadata validates the various bits of data associated
 // with a single attempt.
-func TestAttemptMetadata(t *testing.T) {
+func (s *Suite) TestAttemptMetadata() {
 	sts := SimpleTestSetup{
 		NamespaceName: "TestAttemptMetadata",
 		WorkerName:    "worker",
@@ -143,71 +141,71 @@ func TestAttemptMetadata(t *testing.T) {
 		WorkUnitName:  "a",
 		WorkUnitData:  map[string]interface{}{"from": "wu"},
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
-	attempt := sts.RequestOneAttempt(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
+	attempt := sts.RequestOneAttempt(s)
 
 	// Start checking things
-	start := Clock.Now()
+	start := s.Clock.Now()
 	startTime, err := attempt.StartTime()
-	if assert.NoError(t, err) {
-		assert.WithinDuration(t, start, startTime, 1*time.Millisecond)
+	if s.NoError(err) {
+		s.WithinDuration(start, startTime, 1*time.Millisecond)
 	}
 
-	DataMatches(t, attempt, map[string]interface{}{"from": "wu"})
+	s.DataMatches(attempt, map[string]interface{}{"from": "wu"})
 
 	endTime, err := attempt.EndTime()
-	if assert.NoError(t, err) {
-		assert.WithinDuration(t, time.Time{}, endTime, 1*time.Millisecond)
+	if s.NoError(err) {
+		s.WithinDuration(time.Time{}, endTime, 1*time.Millisecond)
 	}
 
 	expirationTime, err := attempt.ExpirationTime()
-	if assert.NoError(t, err) {
-		assert.Equal(t, 15*time.Minute, expirationTime.Sub(startTime))
+	if s.NoError(err) {
+		s.Equal(15*time.Minute, expirationTime.Sub(startTime))
 	}
 
 	// Renew the lease, giving new work unit data
-	Clock.Add(10 * time.Second)
-	renewTime := Clock.Now()
+	s.Clock.Add(10 * time.Second)
+	renewTime := s.Clock.Now()
 	renewDuration := 5 * time.Minute
 	err = attempt.Renew(renewDuration,
 		map[string]interface{}{"from": "renew"})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	startTime, err = attempt.StartTime()
-	if assert.NoError(t, err) {
+	if s.NoError(err) {
 		// no change from above
-		assert.WithinDuration(t, start, startTime, 1*time.Millisecond)
+		s.WithinDuration(start, startTime, 1*time.Millisecond)
 	}
 
-	DataMatches(t, attempt, map[string]interface{}{"from": "renew"})
+	s.DataMatches(attempt, map[string]interface{}{"from": "renew"})
 
 	endTime, err = attempt.EndTime()
-	if assert.NoError(t, err) {
-		assert.WithinDuration(t, time.Time{}, endTime, 1*time.Millisecond)
+	if s.NoError(err) {
+		s.WithinDuration(time.Time{}, endTime, 1*time.Millisecond)
 	}
 
 	expectedExpiration := renewTime.Add(renewDuration)
 	expirationTime, err = attempt.ExpirationTime()
-	if assert.NoError(t, err) {
-		assert.WithinDuration(t, expectedExpiration, expirationTime, 1*time.Millisecond)
+	if s.NoError(err) {
+		s.WithinDuration(expectedExpiration, expirationTime, 1*time.Millisecond)
 	}
 
 	// Finish the attempt
 	err = attempt.Finish(map[string]interface{}{"from": "finish"})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	startTime, err = attempt.StartTime()
-	if assert.NoError(t, err) {
+	if s.NoError(err) {
 		// no change from above
-		assert.WithinDuration(t, start, startTime, 1*time.Millisecond)
+		s.WithinDuration(start, startTime, 1*time.Millisecond)
 	}
 
-	DataMatches(t, attempt, map[string]interface{}{"from": "finish"})
+	s.DataMatches(attempt, map[string]interface{}{"from": "finish"})
 
 	endTime, err = attempt.EndTime()
-	if assert.NoError(t, err) {
-		assert.True(t, endTime.After(startTime),
+	if s.NoError(err) {
+		s.True(endTime.After(startTime),
 			"start time %+v end time %+v", startTime, endTime)
 	}
 
@@ -216,7 +214,7 @@ func TestAttemptMetadata(t *testing.T) {
 
 // TestWorkUnitChaining tests that completing work units in one work spec
 // will cause work units to appear in another, if so configured.
-func TestWorkUnitChaining(t *testing.T) {
+func (s *Suite) TestWorkUnitChaining() {
 	var (
 		err      error
 		one, two coordinate.WorkSpec
@@ -228,14 +226,14 @@ func TestWorkUnitChaining(t *testing.T) {
 		NamespaceName: "TestWorkUnitChaining",
 		WorkerName:    "worker",
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
 
 	one, err = sts.Namespace.SetWorkSpec(map[string]interface{}{
 		"name": "one",
 		"then": "two",
 	})
-	if !assert.NoError(t, err) {
+	if !s.NoError(err) {
 		return
 	}
 	// RequestAttempts always returns this
@@ -245,61 +243,61 @@ func TestWorkUnitChaining(t *testing.T) {
 		"name":     "two",
 		"disabled": true,
 	})
-	if !assert.NoError(t, err) {
+	if !s.NoError(err) {
 		return
 	}
 
 	// Create and perform a work unit, with no output
 	_, err = one.AddWorkUnit("a", map[string]interface{}{}, coordinate.WorkUnitMeta{})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	sts.WorkSpec = one
-	attempt = sts.RequestOneAttempt(t)
+	attempt = sts.RequestOneAttempt(s)
 	err = attempt.Finish(nil)
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	units, err = two.WorkUnits(coordinate.WorkUnitQuery{})
-	if assert.NoError(t, err) {
-		assert.Empty(t, units)
+	if s.NoError(err) {
+		s.Empty(units)
 	}
 
 	// Create and perform a work unit, with a map output
 	_, err = one.AddWorkUnit("b", map[string]interface{}{}, coordinate.WorkUnitMeta{})
-	assert.NoError(t, err)
+	s.NoError(err)
 
-	attempt = sts.RequestOneAttempt(t)
+	attempt = sts.RequestOneAttempt(s)
 	err = attempt.Finish(map[string]interface{}{
 		"output": map[string]interface{}{
 			"two_b": map[string]interface{}{"k": "v"},
 		},
 	})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	units, err = two.WorkUnits(coordinate.WorkUnitQuery{})
-	if assert.NoError(t, err) {
-		assert.Len(t, units, 1)
-		if assert.Contains(t, units, "two_b") {
-			DataMatches(t, units["two_b"], map[string]interface{}{"k": "v"})
+	if s.NoError(err) {
+		s.Len(units, 1)
+		if s.Contains(units, "two_b") {
+			s.DataMatches(units["two_b"], map[string]interface{}{"k": "v"})
 		}
 	}
 
 	// Create and perform a work unit, with a slice output
 	_, err = one.AddWorkUnit("c", map[string]interface{}{}, coordinate.WorkUnitMeta{})
-	assert.NoError(t, err)
+	s.NoError(err)
 
-	attempt = sts.RequestOneAttempt(t)
+	attempt = sts.RequestOneAttempt(s)
 	err = attempt.Finish(map[string]interface{}{
 		"output": []string{"two_c", "two_cc"},
 	})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	units, err = two.WorkUnits(coordinate.WorkUnitQuery{})
-	if assert.NoError(t, err) {
-		assert.Len(t, units, 3)
-		assert.Contains(t, units, "two_b")
-		assert.Contains(t, units, "two_cc")
-		if assert.Contains(t, units, "two_c") {
-			DataEmpty(t, units["two_c"])
+	if s.NoError(err) {
+		s.Len(units, 3)
+		s.Contains(units, "two_b")
+		s.Contains(units, "two_cc")
+		if s.Contains(units, "two_c") {
+			s.DataEmpty(units["two_c"])
 		}
 	}
 
@@ -307,24 +305,24 @@ func TestWorkUnitChaining(t *testing.T) {
 	_, err = one.AddWorkUnit("d", map[string]interface{}{
 		"output": []string{"two_d"},
 	}, coordinate.WorkUnitMeta{})
-	assert.NoError(t, err)
-	attempt = sts.RequestOneAttempt(t)
+	s.NoError(err)
+	attempt = sts.RequestOneAttempt(s)
 	err = attempt.Finish(nil)
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	units, err = two.WorkUnits(coordinate.WorkUnitQuery{})
-	if assert.NoError(t, err) {
-		assert.Len(t, units, 4)
-		assert.Contains(t, units, "two_b")
-		assert.Contains(t, units, "two_c")
-		assert.Contains(t, units, "two_cc")
-		assert.Contains(t, units, "two_d")
+	if s.NoError(err) {
+		s.Len(units, 4)
+		s.Contains(units, "two_b")
+		s.Contains(units, "two_c")
+		s.Contains(units, "two_cc")
+		s.Contains(units, "two_d")
 	}
 }
 
 // TestChainingMixed uses a combination of strings and tuples in its
 // "output" data.
-func TestChainingMixed(t *testing.T) {
+func (s *Suite) TestChainingMixed() {
 	var (
 		one, two coordinate.WorkSpec
 		attempt  coordinate.Attempt
@@ -336,29 +334,29 @@ func TestChainingMixed(t *testing.T) {
 		NamespaceName: "TestChainingMixed",
 		WorkerName:    "worker",
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
 
 	one, err = sts.Namespace.SetWorkSpec(map[string]interface{}{
 		"name": "one",
 		"then": "two",
 	})
-	if !assert.NoError(t, err) {
+	if !s.NoError(err) {
 		return
 	}
 
 	two, err = sts.Namespace.SetWorkSpec(map[string]interface{}{
 		"name": "two",
 	})
-	if !assert.NoError(t, err) {
+	if !s.NoError(err) {
 		return
 	}
 
 	_, err = one.AddWorkUnit("a", map[string]interface{}{}, coordinate.WorkUnitMeta{})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	sts.WorkSpec = one
-	attempt = sts.RequestOneAttempt(t)
+	attempt = sts.RequestOneAttempt(s)
 	err = attempt.Finish(map[string]interface{}{
 		"output": []interface{}{
 			"key",
@@ -373,20 +371,20 @@ func TestChainingMixed(t *testing.T) {
 			}},
 		},
 	})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	units, err = two.WorkUnits(coordinate.WorkUnitQuery{})
-	if assert.NoError(t, err) {
-		if assert.Contains(t, units, "key") {
-			DataMatches(t, units["key"], map[string]interface{}{"data": "x"})
-			UnitHasPriority(t, units["key"], 10.0)
+	if s.NoError(err) {
+		if s.Contains(units, "key") {
+			s.DataMatches(units["key"], map[string]interface{}{"data": "x"})
+			s.UnitHasPriority(units["key"], 10.0)
 		}
 	}
 }
 
 // TestChainingTwoStep separately renews an attempt to insert an output
 // key, then finishes the work unit; it should still chain.
-func TestChainingTwoStep(t *testing.T) {
+func (s *Suite) TestChainingTwoStep() {
 	var (
 		one, two coordinate.WorkSpec
 		attempt  coordinate.Attempt
@@ -399,29 +397,29 @@ func TestChainingTwoStep(t *testing.T) {
 		NamespaceName: "TestChainingTwoStep",
 		WorkerName:    "worker",
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
 
 	one, err = sts.Namespace.SetWorkSpec(map[string]interface{}{
 		"name": "one",
 		"then": "two",
 	})
-	if !assert.NoError(t, err) {
+	if !s.NoError(err) {
 		return
 	}
 
 	two, err = sts.Namespace.SetWorkSpec(map[string]interface{}{
 		"name": "two",
 	})
-	if !assert.NoError(t, err) {
+	if !s.NoError(err) {
 		return
 	}
 
 	_, err = one.AddWorkUnit("a", map[string]interface{}{}, coordinate.WorkUnitMeta{})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	sts.WorkSpec = one
-	attempt = sts.RequestOneAttempt(t)
+	attempt = sts.RequestOneAttempt(s)
 	err = attempt.Renew(900*time.Second,
 		map[string]interface{}{
 			"output": []interface{}{
@@ -435,17 +433,17 @@ func TestChainingTwoStep(t *testing.T) {
 				}},
 			},
 		})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	err = attempt.Finish(nil)
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	units, err = two.WorkUnits(coordinate.WorkUnitQuery{})
-	if assert.NoError(t, err) {
-		if assert.Contains(t, units, "\x01\x02\x03\x04") {
+	if s.NoError(err) {
+		if s.Contains(units, "\x01\x02\x03\x04") {
 			unit = units["\x01\x02\x03\x04"]
-			DataEmpty(t, unit)
-			UnitHasPriority(t, unit, 0.0)
+			s.DataEmpty(unit)
+			s.UnitHasPriority(unit, 0.0)
 		}
 	}
 }
@@ -453,7 +451,7 @@ func TestChainingTwoStep(t *testing.T) {
 // TestChainingExpiry tests that, if an attempt finishes but is no
 // longer the active attempt, then its successor work units will not
 // be created.
-func TestChainingExpiry(t *testing.T) {
+func (s *Suite) TestChainingExpiry() {
 	var (
 		one, two coordinate.WorkSpec
 		err      error
@@ -464,14 +462,14 @@ func TestChainingExpiry(t *testing.T) {
 		NamespaceName: "TestChainingExpiry",
 		WorkerName:    "worker",
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
 
 	one, err = sts.Namespace.SetWorkSpec(map[string]interface{}{
 		"name": "one",
 		"then": "two",
 	})
-	if !assert.NoError(t, err) {
+	if !s.NoError(err) {
 		return
 	}
 	sts.WorkSpec = one
@@ -480,42 +478,42 @@ func TestChainingExpiry(t *testing.T) {
 		"name":     "two",
 		"disabled": true,
 	})
-	if !assert.NoError(t, err) {
+	if !s.NoError(err) {
 		return
 	}
 
 	// Create and perform a work unit, with no output
 	unit, err = one.AddWorkUnit("a", map[string]interface{}{}, coordinate.WorkUnitMeta{})
-	if !assert.NoError(t, err) {
+	if !s.NoError(err) {
 		return
 	}
 
-	attempt := sts.RequestOneAttempt(t)
+	attempt := sts.RequestOneAttempt(s)
 
 	// But wait!  We got preempted
 	err = unit.ClearActiveAttempt()
-	assert.NoError(t, err)
-	sts.RequestOneAttempt(t)
+	s.NoError(err)
+	sts.RequestOneAttempt(s)
 
 	// Now, let the original attempt finish, trying to generate
 	// more outputs
 	err = attempt.Finish(map[string]interface{}{
 		"output": []string{"unit"},
 	})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	// Since attempt is no longer active, this shouldn't generate
 	// new outputs
 	units, err := two.WorkUnits(coordinate.WorkUnitQuery{})
-	if assert.NoError(t, err) {
-		assert.Empty(t, units)
+	if s.NoError(err) {
+		s.Empty(units)
 	}
 }
 
 // TestChainingDuplicate tests that work unit chaining still works
 // even when the same output work unit is generated twice (it should
 // get retried).
-func TestChainingDuplicate(t *testing.T) {
+func (s *Suite) TestChainingDuplicate() {
 	var (
 		err      error
 		one, two coordinate.WorkSpec
@@ -526,15 +524,15 @@ func TestChainingDuplicate(t *testing.T) {
 		NamespaceName: "TestChainingDuplicate",
 		WorkerName:    "worker",
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
 
 	one, err = sts.Namespace.SetWorkSpec(map[string]interface{}{
 		"name":     "one",
 		"then":     "two",
 		"priority": 1,
 	})
-	if !assert.NoError(t, err) {
+	if !s.NoError(err) {
 		return
 	}
 
@@ -542,140 +540,140 @@ func TestChainingDuplicate(t *testing.T) {
 		"name":     "two",
 		"priority": 2,
 	})
-	if !assert.NoError(t, err) {
+	if !s.NoError(err) {
 		return
 	}
 
 	_, err = one.AddWorkUnit("a", map[string]interface{}{}, coordinate.WorkUnitMeta{})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	_, err = one.AddWorkUnit("b", map[string]interface{}{}, coordinate.WorkUnitMeta{})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	sts.WorkSpec = one
-	attempt = sts.RequestOneAttempt(t)
-	assert.Equal(t, "a", attempt.WorkUnit().Name())
+	attempt = sts.RequestOneAttempt(s)
+	s.Equal("a", attempt.WorkUnit().Name())
 
 	err = attempt.Finish(map[string]interface{}{
 		"output": []string{"z"},
 	})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	sts.WorkSpec = two
-	attempt = sts.RequestOneAttempt(t)
-	assert.Equal(t, "z", attempt.WorkUnit().Name())
+	attempt = sts.RequestOneAttempt(s)
+	s.Equal("z", attempt.WorkUnit().Name())
 
 	err = attempt.Finish(map[string]interface{}{})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	sts.WorkSpec = one
-	attempt = sts.RequestOneAttempt(t)
-	assert.Equal(t, "b", attempt.WorkUnit().Name())
+	attempt = sts.RequestOneAttempt(s)
+	s.Equal("b", attempt.WorkUnit().Name())
 
 	err = attempt.Finish(map[string]interface{}{
 		"output": []string{"z"},
 	})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	sts.WorkSpec = two
-	attempt = sts.RequestOneAttempt(t)
-	assert.Equal(t, "z", attempt.WorkUnit().Name())
+	attempt = sts.RequestOneAttempt(s)
+	s.Equal("z", attempt.WorkUnit().Name())
 
 	err = attempt.Finish(map[string]interface{}{})
-	assert.NoError(t, err)
+	s.NoError(err)
 
-	sts.RequestNoAttempts(t)
+	sts.RequestNoAttempts(s)
 }
 
 // TestAttemptExpiration validates that an attempt's status will switch
 // (on its own) to "expired" after a timeout.
-func TestAttemptExpiration(t *testing.T) {
+func (s *Suite) TestAttemptExpiration() {
 	sts := SimpleTestSetup{
 		NamespaceName: "TestAttemptExpiration",
 		WorkerName:    "worker",
 		WorkSpecName:  "spec",
 		WorkUnitName:  "a",
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
 
 	attempts, err := sts.Worker.RequestAttempts(coordinate.AttemptRequest{})
-	if !(assert.NoError(t, err) && assert.Len(t, attempts, 1)) {
+	if !(s.NoError(err) && s.Len(attempts, 1)) {
 		return
 	}
 	attempt := attempts[0]
 
 	status, err := attempt.Status()
-	if assert.NoError(t, err) {
-		assert.Equal(t, coordinate.Pending, status)
+	if s.NoError(err) {
+		s.Equal(coordinate.Pending, status)
 	}
 
-	sts.RequestNoAttempts(t)
+	sts.RequestNoAttempts(s)
 
 	// There is a default expiration of 15 minutes (checked elsewhere)
 	// So if we wait for, say, 20 minutes we should become expired
-	Clock.Add(time.Duration(20) * time.Minute)
+	s.Clock.Add(time.Duration(20) * time.Minute)
 	status, err = attempt.Status()
-	if assert.NoError(t, err) {
-		assert.Equal(t, coordinate.Expired, status)
+	if s.NoError(err) {
+		s.Equal(coordinate.Expired, status)
 	}
 
 	// The work unit should be "available" for all purposes
 	meta, err := sts.WorkSpec.Meta(true)
-	if assert.NoError(t, err) {
-		assert.Equal(t, 1, meta.AvailableCount)
-		assert.Equal(t, 0, meta.PendingCount)
+	if s.NoError(err) {
+		s.Equal(1, meta.AvailableCount)
+		s.Equal(0, meta.PendingCount)
 	}
-	sts.CheckUnitStatus(t, coordinate.AvailableUnit)
+	sts.CheckUnitStatus(s, coordinate.AvailableUnit)
 
 	// If we request more attempts we should get back the expired
 	// unit again
-	attempt = sts.RequestOneAttempt(t)
-	AttemptStatus(t, coordinate.Pending, attempt)
+	attempt = sts.RequestOneAttempt(s)
+	s.AttemptStatus(coordinate.Pending, attempt)
 }
 
 // TestRetryDelay verifies that the delay option on the Retry() call works.
-func TestRetryDelay(t *testing.T) {
+func (s *Suite) TestRetryDelay() {
 	sts := SimpleTestSetup{
 		NamespaceName: "TestRetryDelay",
 		WorkerName:    "worker",
 		WorkSpecName:  "spec",
 		WorkUnitName:  "unit",
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
 
-	sts.CheckUnitStatus(t, coordinate.AvailableUnit)
+	sts.CheckUnitStatus(s, coordinate.AvailableUnit)
 
-	attempt := sts.RequestOneAttempt(t)
+	attempt := sts.RequestOneAttempt(s)
 	err := attempt.Retry(nil, 90*time.Second)
-	assert.NoError(t, err)
+	s.NoError(err)
 
-	Clock.Add(60 * time.Second)
-	sts.CheckUnitStatus(t, coordinate.DelayedUnit)
-	sts.RequestNoAttempts(t)
+	s.Clock.Add(60 * time.Second)
+	sts.CheckUnitStatus(s, coordinate.DelayedUnit)
+	sts.RequestNoAttempts(s)
 
-	Clock.Add(60 * time.Second)
-	sts.CheckUnitStatus(t, coordinate.AvailableUnit)
-	sts.CheckWorkUnitOrder(t, "unit")
+	s.Clock.Add(60 * time.Second)
+	sts.CheckUnitStatus(s, coordinate.AvailableUnit)
+	sts.CheckWorkUnitOrder(s, "unit")
 }
 
 // TestAttemptFractionalStart verifies that an attempt that starts at
 // a non-integral time (as most of them are) can find itself.  This is
 // a regression test for a specific issue in restclient.
-func TestAttemptFractionalStart(t *testing.T) {
+func (s *Suite) TestAttemptFractionalStart() {
 	sts := SimpleTestSetup{
 		NamespaceName: "TestAttemptFractionalStart",
 		WorkerName:    "worker",
 		WorkSpecName:  "spec",
 		WorkUnitName:  "unit",
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
 
-	Clock.Add(500 * time.Millisecond)
+	s.Clock.Add(500 * time.Millisecond)
 
-	attempt := sts.RequestOneAttempt(t)
+	attempt := sts.RequestOneAttempt(s)
 	// Since restclient uses the start time as one of its keys to
 	// look up attempts, but the code uses two different ways to
 	// format it, having a fractional second means that this call
@@ -684,42 +682,42 @@ func TestAttemptFractionalStart(t *testing.T) {
 	// a lot of tests without the bug fix since *none* of them can
 	// find their own attempts.
 	err := attempt.Finish(nil)
-	assert.NoError(t, err)
+	s.NoError(err)
 }
 
 // TestAttemptGone verifies that, if a work unit is deleted, its
 // attempts return ErrGone for things.
-func TestAttemptGone(t *testing.T) {
+func (s *Suite) TestAttemptGone() {
 	sts := SimpleTestSetup{
 		NamespaceName: "TestAttemptGone",
 		WorkerName:    "worker",
 		WorkSpecName:  "spec",
 		WorkUnitName:  "unit",
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
 
-	attempt := sts.RequestOneAttempt(t)
+	attempt := sts.RequestOneAttempt(s)
 
 	_, err := sts.WorkSpec.DeleteWorkUnits(coordinate.WorkUnitQuery{})
-	assert.NoError(t, err)
+	s.NoError(err)
 
 	_, err = attempt.Status()
 	if err == coordinate.ErrGone {
 		// okay
 	} else if nswu, ok := err.(coordinate.ErrNoSuchWorkUnit); ok {
-		assert.Equal(t, sts.WorkUnitName, nswu.Name)
+		s.Equal(sts.WorkUnitName, nswu.Name)
 	} else if nsws, ok := err.(coordinate.ErrNoSuchWorkSpec); ok {
-		assert.Equal(t, sts.WorkSpecName, nsws.Name)
+		s.Equal(sts.WorkSpecName, nsws.Name)
 	} else {
-		assert.Fail(t, "unexpected error deleting work spec",
+		s.Fail("unexpected error deleting work spec",
 			"%+v", err)
 	}
 }
 
 // TestMaxRetries is a simple test for the max_retries work spec
 // option.
-func TestMaxRetries(t *testing.T) {
+func (s *Suite) TestMaxRetries() {
 	sts := SimpleTestSetup{
 		NamespaceName: "TestMaxRetries",
 		WorkerName:    "worker",
@@ -729,29 +727,29 @@ func TestMaxRetries(t *testing.T) {
 		},
 		WorkUnitName: "unit",
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
 
-	sts.RequestOneAttempt(t)
+	sts.RequestOneAttempt(s)
 
 	// Force the clock far enough ahead that the attempt has expired.
 	// We should get an attempt for the same work unit again.
-	Clock.Add(1 * time.Hour)
-	sts.RequestOneAttempt(t)
+	s.Clock.Add(1 * time.Hour)
+	sts.RequestOneAttempt(s)
 
 	// Force the clock ahead again, and get an attempt.  Since
 	// we exceed max_retries we should not get an attempt.
-	Clock.Add(1 * time.Hour)
-	sts.RequestNoAttempts(t)
-	sts.CheckUnitStatus(t, coordinate.FailedUnit)
+	s.Clock.Add(1 * time.Hour)
+	sts.RequestNoAttempts(s)
+	sts.CheckUnitStatus(s, coordinate.FailedUnit)
 	data, err := sts.WorkUnit.Data()
-	if assert.NoError(t, err) {
-		assert.Equal(t, "too many retries", data["traceback"])
+	if s.NoError(err) {
+		s.Equal("too many retries", data["traceback"])
 	}
 }
 
 // TestMaxRetriesMulti tests both setting max_retries and max_getwork.
-func TestMaxRetriesMulti(t *testing.T) {
+func (s *Suite) TestMaxRetriesMulti() {
 	sts := SimpleTestSetup{
 		NamespaceName: "TestMaxRetries",
 		WorkerName:    "worker",
@@ -761,12 +759,12 @@ func TestMaxRetriesMulti(t *testing.T) {
 			"max_retries": 1,
 		},
 	}
-	sts.SetUp(t)
-	defer sts.TearDown(t)
+	sts.SetUp(s)
+	defer sts.TearDown(s)
 
 	for _, name := range []string{"a", "b", "c", "d"} {
 		_, err := sts.AddWorkUnit(name)
-		if !assert.NoError(t, err) {
+		if !s.NoError(err) {
 			return
 		}
 	}
@@ -776,28 +774,28 @@ func TestMaxRetriesMulti(t *testing.T) {
 		NumberOfWorkUnits: 10,
 	}
 	attempts, err := sts.Worker.RequestAttempts(req)
-	if assert.NoError(t, err) {
-		if assert.Len(t, attempts, 2) {
-			assert.Equal(t, "a", attempts[0].WorkUnit().Name())
-			assert.Equal(t, "b", attempts[1].WorkUnit().Name())
+	if s.NoError(err) {
+		if s.Len(attempts, 2) {
+			s.Equal("a", attempts[0].WorkUnit().Name())
+			s.Equal("b", attempts[1].WorkUnit().Name())
 		}
 	}
 
 	// Let the first work unit finish and the second time out
-	Clock.Add(1 * time.Minute)
+	s.Clock.Add(1 * time.Minute)
 	err = attempts[0].Finish(nil)
-	assert.NoError(t, err)
+	s.NoError(err)
 
-	Clock.Add(1 * time.Hour)
+	s.Clock.Add(1 * time.Hour)
 
 	// Now get two more work units.  We expect the system to find
 	// b and c, notice that b is expired, and keep looking to return
 	// c and d.
 	attempts, err = sts.Worker.RequestAttempts(req)
-	if assert.NoError(t, err) {
-		if assert.Len(t, attempts, 2) {
-			assert.Equal(t, "c", attempts[0].WorkUnit().Name())
-			assert.Equal(t, "d", attempts[1].WorkUnit().Name())
+	if s.NoError(err) {
+		if s.Len(attempts, 2) {
+			s.Equal("c", attempts[0].WorkUnit().Name())
+			s.Equal("d", attempts[1].WorkUnit().Name())
 		}
 	}
 }