+++ doc/changes.md
@@ -6,6 +6,9 @@ Major Changes
 
 * Add a `max_retries` work spec parameter to limit work unit retries
   ([#6](https://github.com/diffeo/go-coordinate/issues/6))
+* Port the generic test suite to use
+  `github.com/stretchr/testify/suite`, rather than the home-grown
+  `cptest` program in this package.
 
 0.4.2 (4 May 2017)
 -----------